safe_requirements.txt
cryptography==41.0.2
PyYAML==6.0.1